Added exception handlers
Added a check in the multiplyMatrices function to verify if the number of columns in the first matrix (cols1) matches the number of rows in the second matrix (matrix2.size()). If they do not match, we throw an std::invalid_argument exception with an appropriate error message.
In the main function, I added surround the matrix multiplication call with a try-catch block to handle any exceptions thrown during the multiplication process. If an exception occurs, we print the error message to the standard error stream (std::cerr).